Change: `DisplayResultOrAnswerButton` を実装して回答結果か回答ボタンを表示するようにした
