+++ apps/web/src/actions/correctResult.ts
@@ -1,5 +1,6 @@
 'use client';
 
+import type { CorrectResult } from '@repo/types';
 import { trpcClient } from '~/trpc/client';
 import { convertStringsToDates } from '~/utils/convertStringsToDates';
 
@@ -12,5 +13,5 @@ export const findCorrectResultByQuestionId = async ({
     questionId,
   });
 
-  return correctResult ? convertStringsToDates(correctResult) : null;
+  return correctResult ? convertStringsToDates<CorrectResult>(correctResult) : null;
 };
