+++ apps/web/src/app/[questionId]/_components/DisplayResultOrAnswerButton/index.ts
@@ -1,0 +1,1 @@
+export { DisplayResultOrAnswerButton } from './DisplayResultOrAnswerButton';
