+++ apps/web/src/app/[questionId]/page.tsx
@@ -5,6 +5,7 @@ import { BackLink } from '~/components/uiParts/BackLink/BackLink';
 import { AnswerBottomInput } from '../../components/models/post/AnswerBottomInput';
 import { PostList } from '../../components/models/post/PostList';
 import { QuestionInformation } from '../../components/models/question/QuestionInformation';
+import { DisplayResultOrAnswerButton } from './_components/DisplayResultOrAnswerButton';
 
 type Props = { params: Promise<{ questionId: string }> };
 
@@ -20,7 +21,10 @@ export default async function Page({ params }: Props) {
       </Box>
       <Stack rowGap={5}>
         <QuestionInformation question={question} />
-        <PostList question={question} />
+        <Stack rowGap={3} pb='240px'>
+          <PostList question={question} />
+          <DisplayResultOrAnswerButton question={question} />
+        </Stack>
         <AnswerBottomInput questionId={questionId} />
       </Stack>
     </Stack>
