+++ apps/web/src/components/models/correctResult/CorrectResultInformation/index.ts
@@ -1,0 +1,1 @@
+export { CorrectResultInformation } from './CorrectResultInformation';
